fastapi_mysql/requirements.txt
fastapi[all]==0.96.0
asyncmy==0.2.8
asyncpg
sqlalchemy
gunicorn    
prometheus-fastapi-instrumentator==5.9.1
python-logging-loki==0.3.1
opentelemetry-instrumentation-fastapi==0.34b0
opentelemetry-instrumentation-logging==0.34b0
opentelemetry-exporter-otlp==1.13.0
opentelemetry-api==1.13.0
opentelemetry-sdk==1.13.0
opentelemetry-exporter-jaeger-thrift==1.13.0
opentelemetry-instrumentation-asyncpg==0.34b0
opentelemetry.instrumentation.mysql==0.34b0
opentelemetry.instrumentation.pymysql==0.34b0
opentelemetry.instrumentation.sqlalchemy==0.34b0